refactor: ResponseErrorHandler 사용하도록 수정
