+++ module-core/module-client/src/main/kotlin/client/github/GithubClient.kt
@@ -1,6 +1,5 @@
-package client
+package client.github
 
-import com.almumol.ossori.client.GithubClientProperties
 import dto.response.GithubRepositoriesResponse
 import dto.response.GithubRepositoryResponse
 import org.springframework.boot.context.properties.EnableConfigurationProperties
@@ -11,7 +10,7 @@ import org.springframework.web.client.RestClient
 @EnableConfigurationProperties(GithubClientProperties::class)
 @Component
 class GithubClient(
-    private val restClient: RestClient,
+    private val githubRestClient: RestClient,
     private val githubClientProperties: GithubClientProperties
 ) {
 
@@ -35,17 +34,11 @@ class GithubClient(
     }
 
     private inline fun <reified T> defaultGithubGetClient(uri: String): T {
-        val response = restClient.get()
+        val response = githubRestClient.get()
             .uri(uri)
             .accept(APPLICATION_JSON)
             //.header(AUTHORIZATION_HEADER, getGithubToken())
             .retrieve()
-            .onStatus({ it.is4xxClientError }) { _, response ->
-                throw RuntimeException("Status code: ${response.statusCode}, Description: ${response.body}")
-            }
-            .onStatus({ it.is5xxServerError }) { _, _ ->
-                throw RuntimeException("Github Server Unavailable")
-            }
             .body(T::class.java)
         return response ?: throw RuntimeException("Response body is null")
     }
