+++ module-core/module-client/src/main/kotlin/client/github/GithubClientProperties.kt
@@ -1,4 +1,4 @@
-package com.almumol.ossori.client
+package client.github
 
 import org.springframework.boot.context.properties.ConfigurationProperties
