+++ module-core/module-client/src/main/kotlin/global/exception/GithubResponseException.kt
@@ -1,0 +1,5 @@
+package global.exception
+
+class GithubResponseException(message: String) : RuntimeException(message) {
+
+}
